Only inspect correctly scoped declarations in declaration inspection base classes
In the case where the declaration type does not matter all declarations were inspected instead of those in the module. This led to results being generated multiple times if there are multiple modules.

In addition, common behaviour of the declaration inspection base classes has been extracted into a further base class.
